creating review package controller, moving methods, editing routing
