+++ config/routes.rb
@@ -329,15 +329,15 @@ Rails.application.routes.draw do
     post '/correspondence/:correspondence_uuid/correspondence_intake_task', to: 'correspondence_tasks#create_correspondence_intake_task'
     post '/correspondence/:id/remove_package', to: 'correspondence_tasks#remove_package'
     post '/correspondence/:id/completed_package', to: 'correspondence_tasks#completed_package'
-    get '/correspondence/:correspondence_uuid/review_package', to: 'correspondence#review_package'
-    get '/correspondence/edit_document_type_correspondence', to: 'correspondence#document_type_correspondence'
+    get '/correspondence/:correspondence_uuid/review_package', to: 'correspondence_review_package#review_package'
+    get '/correspondence/edit_document_type_correspondence', to: 'correspondence_review_package#document_type_correspondence'
     patch '/correspondence/:correspondence_uuid/intake_update', to: 'correspondence_intake#intake_update'
     get '/correspondence/:correspondence_uuid/veteran', to: 'correspondence#veteran'
-    put '/correspondence/:correspondence_uuid/update_cmp', to: 'correspondence#update_cmp'
-    get '/correspondence/packages', to: 'correspondence#package_documents'
-    get '/correspondence/:correspondence_uuid', to: 'correspondence#show'
-    get '/correspondence/:pdf_id/pdf', to: 'correspondence#pdf'
-    patch '/correspondence/:correspondence_uuid', to: 'correspondence#update'
+    put '/correspondence/:correspondence_uuid/update_cmp', to: 'correspondence_review_package#update_cmp'
+    get '/correspondence/packages', to: 'correspondence_review_package#package_documents'
+    get '/correspondence/:correspondence_uuid', to: 'correspondence_review_package#show'
+    get '/correspondence/:pdf_id/pdf', to: 'correspondence_review_package#pdf'
+    patch '/correspondence/:correspondence_uuid', to: 'correspondence_review_package#update'
     patch '/correspondence/:id/update_document', to: 'correspondence_document#update_document'
     post '/correspondence/:correspondence_uuid', to: 'correspondence_intake#process_intake', as: :queue_correspondence_intake_process_intake
     post "/correspondence/:correspondence_uuid/task", to: "correspondence_tasks#create_package_action_task"
